conti getSearchPagination

+++ src/user/user.service.ts
@@ -2,17 +2,19 @@ import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { PrismaClient } from '@prisma/client'
 import { connect } from 'http2';
 import { UserDto } from './dto/user.dto';
+import { UserSearchPagination } from './dto/userSearchPagination.dto';
 
 @Injectable()
 export class UserService {
 
     private prisma: PrismaClient = new PrismaClient()
 
-    async getUser(): Promise<UserDto[]> {
+    async getUser(): Promise<any> {
         try {
-            return await this.prisma.user.findMany()
+            const data = await this.prisma.user.findMany()
+            return new HttpException({ data, message: "Get user success" }, 200)
         } catch (error) {
-            throw new HttpException({ reason: error }, HttpStatus.BAD_REQUEST)
+            return new HttpException({ reason: error }, HttpStatus.BAD_REQUEST)
         }
     }
 
@@ -22,7 +24,7 @@ export class UserService {
             const checkEmail = await this.prisma.user.findFirst({ where: { email: data.email } })
             if (!checkEmail) {
                 await this.prisma.user.create({ data })
-                return new HttpException(data, 200)
+                return new HttpException({ data, message: "User added" }, 200)
             }
             return new HttpException("Email already used", 401)
         } catch (error) {
@@ -33,10 +35,28 @@ export class UserService {
 
     async deleteUser(user_id: string): Promise<any> {
         try {
-            await this.prisma.user.delete({
+            const checkUserId = await this.prisma.user.findFirst({
                 where: { user_id: Number(user_id) }
             })
-            return HttpStatus.OK
+            if (!checkUserId) return new HttpException("User already deleted or not existed", 401)
+
+            const data = await this.prisma.user.delete({
+                where: { user_id: Number(user_id) }
+            })
+            return new HttpException({ data, message: "User deleted" }, 200)
+        } catch (error) {
+            return new HttpException({ message: error }, HttpStatus.BAD_REQUEST)
+        }
+    }
+
+    //chưa biết làm pageIndex (trang số Index)
+    async getSearchPagination(input: UserSearchPagination): Promise<any> {
+        try {
+            const data = { ...input }
+            const dataSearch = await this.prisma.user.findMany({
+                where: { name: { contains: data.keyword } }, take: data.pageSize
+            })
+            return new HttpException({ dataSearch, message: "Get search pagination success" }, HttpStatus.OK)
         } catch (error) {
             return new HttpException({ message: error }, HttpStatus.BAD_REQUEST)
         }
